Fix: Path equality checks
