+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/etsi/ElectronicAddress.kt
@@ -1,5 +1,6 @@
 package at.asitplus.etsi
 
+import at.asitplus.rfc.Rfc3986UriSchemeName
 import kotlinx.serialization.Serializable
 import kotlin.jvm.JvmInline
