--- etsi-data-classes/src/commonMain/kotlin/at/asitplus/etsi/Rfc3986UriPathEmpty.kt
@@ -1,5 +1,0 @@
-package at.asitplus.etsi
-
-object Rfc3986UriPathEmpty : Rfc3986UriPath, Rfc3986RelativeReferencePath {
-    override fun toString() = ""
-}
