+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/etsi/Rfc5646LanguageTag.kt
@@ -1,12 +1,7 @@
 package at.asitplus.etsi
 
-import kotlinx.serialization.KSerializer
+import at.asitplus.rfc.CaseInsensitiveString
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 import kotlin.jvm.JvmInline
 
 /**
