+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986AuthorityHostIPvFuture.kt
@@ -1,4 +1,4 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
@@ -8,7 +8,6 @@ import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
-import kotlin.jvm.JvmInline
 
 @Serializable(with = Rfc3986AuthorityHostIPvFuture.InlineSerializer::class)
 data class Rfc3986AuthorityHostIPvFuture(
