+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986AuthorityHostRegisteredName.kt
@@ -1,4 +1,4 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
 import kotlinx.serialization.Serializable
 import kotlin.jvm.JvmInline
