+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986Path.kt
@@ -1,4 +1,4 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
 sealed interface Rfc3986Path {
     fun validate() {
@@ -8,5 +8,16 @@ sealed interface Rfc3986Path {
                 "Expected path to consist of pchar and `/`, but got `$ch` at index $index in `$string`."
             }
         }
+    }
+
+    fun equalsPath(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null) return false
+
+        if (other !is Rfc3986Path) return false
+
+        if (Rfc3986UriPathEmpty.toString() != other.toString()) return false
+
+        return true
     }
 }
