+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986RelativeReference.kt
@@ -1,4 +1,4 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
@@ -8,15 +8,15 @@ import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 
-
-@Serializable(with = Rfc3986UniformResourceIdentifier.InlineSerializer::class)
-data class Rfc3986UniformResourceIdentifier(
-    override val schemeName: Rfc3986UriSchemeName,
+@Serializable(with = Rfc3986RelativeReference.InlineSerializer::class)
+data class Rfc3986RelativeReference(
     override val authority: Rfc3986Authority? = null,
-    override val path: Rfc3986UriPath,
+    override val path: Rfc3986RelativeReferencePath,
     override val query: Rfc3986UriQuery? = null,
     override val fragment: Rfc3986UriFragment? = null,
 ): Rfc3986UriReference {
+    override val schemeName: Rfc3986UriSchemeName?
+        get() = null
     init {
         if (authority != null) {
             // All of these are necessary since Rfc3986UriPathAbsoluteOrEmpty allows for the first segment to be empty
@@ -30,7 +30,6 @@ data class Rfc3986UniformResourceIdentifier(
 
     val string: String by lazy {
         listOfNotNull(
-            "$schemeName:",
             authority?.toString(true),
             path.toString(),
             query?.let { "?$it" },
@@ -39,18 +38,17 @@ data class Rfc3986UniformResourceIdentifier(
     }
 
     companion object {
-        operator fun invoke(string: String): Rfc3986UniformResourceIdentifier {
-            val reference = Rfc3986UriReference(string)
-            when(reference) {
-                is Rfc3986UniformResourceIdentifier -> return reference
-                is Rfc3986RelativeReference -> throw IllegalArgumentException(
-                    "Expected URI to contain a scheme, but none was found in `$string`."
+        operator fun invoke(string: String): Rfc3986RelativeReference {
+            when(val reference = Rfc3986UriReference(string)) {
+                is Rfc3986RelativeReference -> return reference
+                is Rfc3986UniformResourceIdentifier -> throw IllegalArgumentException(
+                    "Expected relative reference to contain no scheme, but got `${reference.schemeName}` in `$string`."
                 )
             }
         }
     }
 
-    class InlineSerializer : KSerializer<Rfc3986UniformResourceIdentifier> {
+    class InlineSerializer : KSerializer<Rfc3986RelativeReference> {
         override val descriptor: SerialDescriptor
             get() = PrimitiveSerialDescriptor(
                 serialName = InlineSerializer::class.qualifiedName!!,
@@ -59,14 +57,13 @@ data class Rfc3986UniformResourceIdentifier(
 
         override fun serialize(
             encoder: Encoder,
-            value: Rfc3986UniformResourceIdentifier
+            value: Rfc3986RelativeReference
         ) {
             encoder.encodeString(value.string)
         }
 
-        override fun deserialize(decoder: Decoder) = Rfc3986UniformResourceIdentifier(
+        override fun deserialize(decoder: Decoder) = Rfc3986RelativeReference(
             decoder.decodeString()
         )
     }
 }
-
