+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986UniformResourceIdentifier.kt
@@ -1,4 +1,4 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
@@ -8,15 +8,14 @@ import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 
-@Serializable(with = Rfc3986RelativeReference.InlineSerializer::class)
-data class Rfc3986RelativeReference(
+@Serializable(with = Rfc3986UniformResourceIdentifier.InlineSerializer::class)
+data class Rfc3986UniformResourceIdentifier(
+    override val schemeName: Rfc3986UriSchemeName,
     override val authority: Rfc3986Authority? = null,
-    override val path: Rfc3986RelativeReferencePath,
+    override val path: Rfc3986UriPath,
     override val query: Rfc3986UriQuery? = null,
     override val fragment: Rfc3986UriFragment? = null,
 ): Rfc3986UriReference {
-    override val schemeName: Rfc3986UriSchemeName?
-        get() = null
     init {
         if (authority != null) {
             // All of these are necessary since Rfc3986UriPathAbsoluteOrEmpty allows for the first segment to be empty
@@ -30,6 +29,7 @@ data class Rfc3986RelativeReference(
 
     val string: String by lazy {
         listOfNotNull(
+            "$schemeName:",
             authority?.toString(true),
             path.toString(),
             query?.let { "?$it" },
@@ -38,17 +38,18 @@ data class Rfc3986RelativeReference(
     }
 
     companion object {
-        operator fun invoke(string: String): Rfc3986RelativeReference {
-            when(val reference = Rfc3986UriReference(string)) {
-                is Rfc3986RelativeReference -> return reference
-                is Rfc3986UniformResourceIdentifier -> throw IllegalArgumentException(
-                    "Expected relative reference to contain no scheme, but got `${reference.schemeName}` in `$string`."
+        operator fun invoke(string: String): Rfc3986UniformResourceIdentifier {
+            val reference = Rfc3986UriReference(string)
+            when(reference) {
+                is Rfc3986UniformResourceIdentifier -> return reference
+                is Rfc3986RelativeReference -> throw IllegalArgumentException(
+                    "Expected URI to contain a scheme, but none was found in `$string`."
                 )
             }
         }
     }
 
-    class InlineSerializer : KSerializer<Rfc3986RelativeReference> {
+    class InlineSerializer : KSerializer<Rfc3986UniformResourceIdentifier> {
         override val descriptor: SerialDescriptor
             get() = PrimitiveSerialDescriptor(
                 serialName = InlineSerializer::class.qualifiedName!!,
@@ -57,13 +58,14 @@ data class Rfc3986RelativeReference(
 
         override fun serialize(
             encoder: Encoder,
-            value: Rfc3986RelativeReference
+            value: Rfc3986UniformResourceIdentifier
         ) {
             encoder.encodeString(value.string)
         }
 
-        override fun deserialize(decoder: Decoder) = Rfc3986RelativeReference(
+        override fun deserialize(decoder: Decoder) = Rfc3986UniformResourceIdentifier(
             decoder.decodeString()
         )
     }
 }
+
