+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986UriAuthorityUserInformation.kt
@@ -1,12 +1,6 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 import kotlin.jvm.JvmInline
 
 @Serializable
