+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986UriPath.kt
@@ -1,5 +1,4 @@
-package at.asitplus.etsi
-
+package at.asitplus.rfc
 
 sealed interface Rfc3986UriPath : Rfc3986Path {
     companion object {
