+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986UriPathAbsolute.kt
@@ -1,9 +1,6 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
-import kotlin.jvm.JvmInline
-
-@JvmInline
-value class Rfc3986UriPathAbsolute(
+data class Rfc3986UriPathAbsolute(
     val percentEncodingAwareString: Rfc3986PercentEncodingAwareString
 ) : Rfc3986UriPath, Rfc3986RelativeReferencePath {
     init {
@@ -23,4 +20,8 @@ value class Rfc3986UriPathAbsolute(
         get() = percentEncodingAwareString.string
 
     override fun toString() = string
+
+    override fun equals(other: Any?) = equalsPath(other)
+
+    override fun hashCode() = toString().hashCode()
 }
