+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986UriPathAbsoluteOrEmpty.kt
@@ -1,9 +1,6 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
-import kotlin.jvm.JvmInline
-
-@JvmInline
-value class Rfc3986UriPathAbsoluteOrEmpty(
+data class Rfc3986UriPathAbsoluteOrEmpty(
     val percentEncodingAwareString: Rfc3986PercentEncodingAwareString,
 ) : Rfc3986UriPath, Rfc3986RelativeReferencePath {
     init {
@@ -20,4 +17,8 @@ value class Rfc3986UriPathAbsoluteOrEmpty(
         get() = percentEncodingAwareString.string
 
     override fun toString() = string
+
+    override fun equals(other: Any?) = equalsPath(other)
+
+    override fun hashCode() = toString().hashCode()
 }
