+++ etsi-data-classes/src/commonMain/kotlin/at/asitplus/rfc/Rfc3986UriSchemeName.kt
@@ -1,12 +1,6 @@
-package at.asitplus.etsi
+package at.asitplus.rfc
 
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 import kotlin.jvm.JvmInline
 
 /**
@@ -27,6 +21,13 @@ value class Rfc3986UriSchemeName(
                 "Expected scheme name to consist of letters (a-z, A-Z), digits (0-9), `+`, `-` or `.`, but got `$it` at index $index in `$string`"
             }
         }
+        require(Regex(REGEX).matches(string)) {
+            "Expected scheme name to satisfy regex ^$REGEX\$, but got `$string`."
+        }
+    }
+
+    companion object {
+        const val REGEX = "([a-zA-Z][a-zA-Z0-9+-.]*)"
     }
 
     object Common {
@@ -36,6 +37,7 @@ value class Rfc3986UriSchemeName(
     }
 
     constructor(string: String) : this(CaseInsensitiveString(string))
+
     val string: String
         get() = caseInsensitiveString.string
