+++ etsi-data-classes/src/commonTest/kotlin/at/asitplus/etsi/Rfc3986UriAuthorityHostTest.kt
@@ -1,9 +1,9 @@
 package at.asitplus.etsi
 
+import at.asitplus.rfc.Rfc3986AuthorityHost
 import at.asitplus.testballoon.withData
 import de.infix.testBalloon.framework.core.testSuite
 import io.kotest.assertions.throwables.shouldNotThrowAny
-import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 
 @Suppress("unused")
