+++ etsi-data-classes/src/commonTest/kotlin/at/asitplus/etsi/Rfc3986UriSchemeNameTest.kt
@@ -1,5 +1,6 @@
 package at.asitplus.etsi
 
+import at.asitplus.rfc.Rfc3986UriSchemeName
 import at.asitplus.testballoon.withData
 import de.infix.testBalloon.framework.core.testSuite
 import io.kotest.assertions.throwables.shouldThrow
